requirements.txt
mutagen==1.42.0
Flask==1.1.1
pydub==0.23.1
google-cloud-storage==1.23.0
google-cloud-texttospeech==0.5.0
google-cloud-translate==2.0.0



